requirements.txt
amadeus==12.0.0
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
bleach==6.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
Django==6.0
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
geographiclib==2.1
geopy==2.4.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.14.0
packaging==25.0
pip_system_certs==5.3
pydantic==2.12.5
pydantic_core==2.41.5
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-xdist==3.5.0
pytest-randomly==3.16.0
redis==7.1.0
requests==2.32.5
responses==0.25.6
sniffio==1.3.1
sqlparse==0.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
webencodings==0.5.1
